Renamed folders in models and tests as "issues"
